+++ filter/regexp.go
@@ -31,9 +31,11 @@ import (
 //
 // FilterExpression defines the regular expression used for matching the message
 // payload. If the expression matches, the message is passed.
+// FilterExpression is evaluated after FilterExpressionNot.
 //
 // FilterExpressionNot defines a negated regular expression used for matching
 // the message payload. If the expression matches, the message is blocked.
+// FilterExpressionNot is evaluated before FilterExpression.
 type RegExp struct {
 	exp    *regexp.Regexp
 	expNot *regexp.Regexp
@@ -55,9 +57,9 @@ func (filter *RegExp) Configure(conf core.PluginConfig) error {
 		}
 	}
 
-	exp = conf.GetString("FilterExpressionNot", "")
-	if exp != "" {
-		filter.expNot, err = regexp.Compile(exp)
+	notExp := conf.GetString("FilterExpressionNot", "")
+	if notExp != "" {
+		filter.expNot, err = regexp.Compile(notExp)
 		if err != nil {
 			return err // ### return, regex parser error ###
 		}
@@ -68,12 +70,14 @@ func (filter *RegExp) Configure(conf core.PluginConfig) error {
 
 // Accepts allows all messages
 func (filter *RegExp) Accepts(msg core.Message) bool {
-	if filter.exp != nil {
-		return filter.exp.MatchString(string(msg.Data))
+	if filter.expNot != nil {
+		if filter.expNot.MatchString(string(msg.Data)) {
+			return false
+		}
 	}
 
-	if filter.expNot != nil {
-		return !filter.expNot.MatchString(string(msg.Data))
+	if filter.exp != nil {
+		return filter.exp.MatchString(string(msg.Data))
 	}
 
 	return true // ### return, pass everything ###